Clean encode context before putting to pool
Currently, once encode runtime context is done, we put it back to pool
to reuse later. The unsafe pointers in context keepRefs are also kept alive,
but its underlying object maybe cleaned already. So if the garbage
collector run, it will see an invalid pointer, the runtime crashes.

To fix it, we add encodeRuntimeContext.done method, to clean the keepRefs
once we finish using the context.

Fixes #63